requirements.txt
python-telegram-bot>=20.0
python-dotenv
apscheduler
gtts
openai
anthropic
aiohttp
aiofiles